Test file has been moved as requested
